// src/app/models/srm/providertimeline.ts
export class Providertimeline {
    idProductxSupplier:number= -1;
    dateCreate: Date;
    dateUpdate: Date;
    idReasons:number=-1;
    dateofOcurrence: Date;
    description:string="";
    createdByUser: string="";
    updatedByUser: string="";
    // statusId: number = -1;
    typeOccurrence: string="";
    //adicionales
    icon: string="";
    color: string="";
    image: string="";
}